update gatsby example project
