+++ examples/gatsby/package.json
@@ -20,8 +20,11 @@
     "@orgajs/loader": "3.1.2",
     "@orgajs/react": "3.0.1",
     "gatsby": "^3.10.0",
+    "gatsby-plugin-catch-links": "^3.14.0",
+    "gatsby-plugin-image": "^1.14.0",
     "gatsby-plugin-orga": "3.1.3",
-    "gatsby-source-filesystem": "^3.3.0",
+    "gatsby-plugin-sharp": "^3.14.0",
+    "gatsby-source-filesystem": "^3.14.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1"
   }
