+++ examples/gatsby/src/pages/index.org
@@ -1,7 +1,9 @@
 #+TITLE: Orga + Gatsby = Awesome
 #+jsx: import Box from '../components/box'
-#+jsx: import Layout from '../components/layout'
-#+jsx: export default Layout
+#+layout: ../components/layout
+
+#+attr_html: :width 300
+[[file:../images/logo.png][orgajs]]
 
 * Hi
 You can /embed/ react components like [[https://mdxjs.com/][mdx]], but better, with =export jsx block=.
@@ -9,6 +11,9 @@ You can /embed/ react components like [[https://mdxjs.com/][mdx]], but better, w
 #+begin_export jsx
 <Box>that tomato box</Box>
 #+end_export
+
+- [[file:page2.org][here is a link to another org file]]
+- [[file:page3.tsx][here is a link to a tsx file]]
 
 #+begin_export jsx
 export const query = graphql`
