+++ examples/gatsby/src/pages/page2.org
@@ -1,0 +1,3 @@
+#+TITLE: Page2
+
+This is another page generated from org-mode file.
